Runtime: 112 ms, faster than 75.35% of C online submissions for Palindrome Number. Memory Usage: 72.2 MB, less than 6.38% of C online submissions for Palindrome Number. 9. [Palindrome Number](https://leetcode.com/problems/palindrome-number/)
Determine whether an integer is a palindrome. An integer is a palindrome 
when it reads the same backward as forward.

Example 1:
```
Input: 121
Output: true
```

Example 2:
```
Input: -121
Output: false

Explanation: From left to right, it reads -121. From right to left, it 
becomes 121-. Therefore it is not a palindrome.
```
```
Example 3:

Input: 10
Output: false
Explanation: Reads 01 from right to left. Therefore it is not a 
palindrome.
```


Coud you solve it without converting the integer to a string?